+++ src/api/notice/NoticeApi.js
@@ -125,11 +125,28 @@ export const NoticeApi = {
 
 
   // 파일 다운로드
-  downloadFile: async (noticeId, fileId) => {
+  // NoticeApi.js
+  downloadFile: async (noticeId, fileId, fileName) => {
     try {
       const response = await apiClient.get(`/api/notices/${noticeId}/files/${fileId}/download`, {
-        responseType: 'blob'
+        responseType: 'blob'  // 중요: responseType을 'blob'으로 설정
       });
+
+      // Blob URL 생성
+      const blob = new Blob([response.data]);
+      const url = window.URL.createObjectURL(blob);
+
+      // 임시 링크 생성 및 클릭
+      const link = document.createElement('a');
+      link.href = url;
+      link.download = fileName; // 서버에서 전송된 파일명 사용
+      document.body.appendChild(link);
+      link.click();
+
+      // cleanup
+      document.body.removeChild(link);
+      window.URL.revokeObjectURL(url);
+
       return response.data;
     } catch (error) {
       console.error('파일 다운로드 실패:', error);
@@ -137,5 +154,6 @@ export const NoticeApi = {
     }
   },
 };
+
 
 export default NoticeApi;
